pyproject.toml
[tool.poetry]
name = "telebotnew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.10"
newsapi = "^0.1.1"
python-newsapi = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
